requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
pandas>=1.5.0
pyarrow>=8.0.0